// src/payload-types.ts
/* tslint:disable */
/* eslint-disable */
/**
 * This file was automatically generated by Payload.
 * DO NOT MODIFY IT BY HAND. Instead, modify your source Payload config,
 * and re-run `payload generate:types` to regenerate this file.
 */

export interface Config {
  collections: {
    users: User
    products: Product
    media: Media
    product_files: ProductFile
    orders: Order
    "payload-preferences": PayloadPreference
    "payload-migrations": PayloadMigration
  }
  globals: {}
}
export interface User {
  id: string
  products?: (string | Product)[] | null
  product_files?: (string | ProductFile)[] | null
  role: "admin" | "user"
  updatedAt: string
  createdAt: string
  email: string
  resetPasswordToken?: string | null
  resetPasswordExpiration?: string | null
  salt?: string | null
  hash?: string | null
  _verified?: boolean | null
  _verificationToken?: string | null
  loginAttempts?: number | null
  lockUntil?: string | null
  password: string | null
}
export interface Product {
  id: string
  user?: (string | null) | User
  name: string
  description?: string | null
  price: number
  category: "ui_kits" | "icons"
  product_files: string | ProductFile
  approvedForSale?: ("pending" | "approved" | "denied") | null
  priceId?: string | null
  stripeId?: string | null
  images: {
    image: string | Media
    id?: string | null
  }[]
  updatedAt: string
  createdAt: string
}
export interface ProductFile {
  id: string
  user?: (string | null) | User
  updatedAt: string
  createdAt: string
  url?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
}
export interface Media {
  id: string
  user?: (string | null) | User
  updatedAt: string
  createdAt: string
  url?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
  sizes?: {
    thumbnail?: {
      url?: string | null
      width?: number | null
      height?: number | null
      mimeType?: string | null
      filesize?: number | null
      filename?: string | null
    }
    card?: {
      url?: string | null
      width?: number | null
      height?: number | null
      mimeType?: string | null
      filesize?: number | null
      filename?: string | null
    }
    tablet?: {
      url?: string | null
      width?: number | null
      height?: number | null
      mimeType?: string | null
      filesize?: number | null
      filename?: string | null
    }
  }
}
export interface Order {
  id: string
  _isPaid: boolean
  user: string | User
  products: (string | Product)[]
  updatedAt: string
  createdAt: string
}
export interface PayloadPreference {
  id: string
  user: {
    relationTo: "users"
    value: string | User
  }
  key?: string | null
  value?:
    | {
        [k: string]: unknown
      }
    | unknown[]
    | string
    | number
    | boolean
    | null
  updatedAt: string
  createdAt: string
}
export interface PayloadMigration {
  id: string
  name?: string | null
  batch?: number | null
  updatedAt: string
  createdAt: string
}

declare module "payload" {
  export interface GeneratedTypes extends Config {}
}